% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{apply_duplicate_suffixes}
\alias{apply_duplicate_suffixes}
\title{Apply duplicate suffixes to tips and nodes}
\usage{
apply_duplicate_suffixes(phy)
}
\arguments{
\item{phy}{A phylo object, the tree whose labels are to have copy suffixes
applied.}
}
\value{
A phylo object, the same tree but with suffixes applied to the
  labels.
}
\description{
Suffixes are applied to ensure tip labels and node labels are not duplicates.
Suffixes have the form -1, -2, -3, ...
}
\details{
The function recognizes existing duplicate suffixes and deals with them in
one of two ways. If a label has n duplicates that are already suffixed -1,
-2, ... -n, then the suffixes are not changed. Under any other conditions,
old suffixes are removed and new ones applied.

Suffixation of tips and of nodes are handled independently of one another.
}
\examples{

library(ape)

tree <- abridge_labels(get_glottolog_trees("Koreanic"))
plot(tree)
nodelabels(tree$node.label)
tree2 <- clone_tip(tree, "chol1278", n = 2, subgroup = TRUE)
plot(tree2)
nodelabels(tree2$node.label)
# Technically, tree2 is ill-formed because it has duplicate tip labels.
# Note how this causes problems if we try to clone one of them, since it
# is unclear which should be cloned:
\dontrun{
tree2a <- clone_tip(tree2, "chol1278")
}

# Suffixation of duplicate tips
tree3 <- apply_duplicate_suffixes(tree2)
plot(tree3)
nodelabels(tree3$node.label)

# Once they are suffixed, these tips can be cloned successfully:
tree4 <- clone_tip(tree3, c("chol1278-2", "chol1278-3"), subgroup = TRUE)
plot(tree4)
nodelabels(tree4$node.label)
# Suffixing is applied across all tips that share a glottocode, and
# separately, across all nodes that share a glottocode:
tree5 <- apply_duplicate_suffixes(tree4)
plot(tree5)
nodelabels(tree5$node.label)
}
