% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{assemble_rake}
\alias{assemble_rake}
\title{Bind trees as a high-level rake}
\usage{
assemble_rake(phy)
}
\arguments{
\item{phy}{A multiphylo object containing the trees to be combined.}
}
\value{
A phylo object, a single tree.
}
\description{
Takes a multiPhylo object containing multiple trees and combines them into a
single tree with a rake structure at its root, below which each tree appears
on its own branch.
}
\examples{

library(ape)

arnhem_hypothesis <- 
  c("Gunwinyguan", "Mangarrayi-Maran", "Maningrida",
    "Kungarakany", "Gaagudju")
trees <- get_glottolog_trees(arnhem_hypothesis)
simple_rake <- assemble_rake(trees)
rake_in_rake <- assemble_rake(c(assemble_rake(trees[1:3]), trees[4:5]))
plot(simple_rake)
plot(rake_in_rake)

# If `phy` contains only one tree, a warning is issued.
mono_rake <- assemble_rake(trees[3])
plot(mono_rake)
}
