% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{collapse_node}
\alias{collapse_node}
\title{Collapse one or more nodes rootwards}
\usage{
collapse_node(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character vector containing node labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
Collapse one or more nodes rootwards
}
\examples{

library(ape)
tree <- 
  rescale_branches_exp(abridge_labels(get_glottolog_trees("Tangkic")))
plot(tree)
nodelabels(tree$node.label)
tree2 <- collapse_node(tree, "gang1267")
plot(tree2)
nodelabels(tree2$node.label)
}
