% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{convert_to_tip}
\alias{convert_to_tip}
\title{Convert nodes to tips}
\usage{
convert_to_tip(phy, label, warn = TRUE)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character vector containing node or tip labels.}

\item{warn}{A logical, whether to issue warning when \code{label} cantains
tip labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
For nodes identified by their label, remove the clade that they dominate and
replace it with a tip of the same name.
}
\details{
Any labels included in \code{label} which are both node labels and tip labels
are regarded as node label, and that node will be removed and replaced by a
tip.

Any labels included in \code{label} which are tip labels only are ignored.
}
