% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{extract_glottocode}
\alias{extract_glottocode}
\title{Extract glottocode substrings}
\usage{
extract_glottocode(labels)
}
\arguments{
\item{label}{A string}
}
\value{
A string
}
\description{
From a character vector, extracts the first glottocode from each element.
}
\details{
Glottocodes comprise four lowercase letters (or b10b or 3adt) followed by
four numbers, and are only identified if they are initial in the string or
are preceded by [.

Also recognizes and retains duplicate suffixes, i.e., a hyphen followed by
one or more numerals at the end of the string (see
\link{apply_duplicate_suffixes}).
}
\examples{

extract_glottocode("DongoKresh[dong1296]-l-")
extract_glottocode(c("DongoKresh[dong1296]-l-", "Goro-Golo[orlo1238]"))

# Duplicate suffixes are recognised and retained
extract_glottocode(c("Goro-Golo[orlo1238]-1", "Goro-Golo[orlo1238]-2"))
}
