% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_glottolog_languages}
\alias{get_glottolog_languages}
\title{Simple language metadata}
\usage{
get_glottolog_languages(glottolog_version)
}
\arguments{
\item{glottolog_version}{A character string, specifying which glottolog
version to use. Currently available options are \code{'4.3'} and
\code{'4.4'}. If no value is specified then the newest available version is
used.}
}
\description{
Returns a dataframe of metadata on glottolog's languages.
}
\details{
Returned columns are: \code{glottocode}, \code{isocodes}, \code{name},
\code{name_in_tree}, \code{position}, \code{tree} and \code{tree_name}.
}
\examples{
head(get_glottolog_languages())
head(get_glottolog_languages(glottolog_version = "4.3"))
}
