% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_glottolog_phylo_geo}
\alias{get_glottolog_phylo_geo}
\title{Extended glottolog metadata}
\usage{
get_glottolog_phylo_geo(glottolog_version)
}
\arguments{
\item{glottolog_version}{A character string, specifying which glottolog
version to use. Currently available options are \code{'4.3'} and
\code{'4.4'}. If no value is specified then the newest available version is
used.}
}
\description{
Returns a dataframe of glottolog geographical and phylogenetic metadata.
}
\details{
Returned columns are: \code{glottocode}, \code{isocodes}, \code{name}, \code{level},
\code{vertex_type}, \code{vertex_label}, \code{vertex_name},
\code{macroarea}, \code{latitude}, \code{longitude},
\code{family_glottocode}, \code{family_name} and \code{tree}.
}
\examples{
head(get_glottolog_phylo_geo())
head(get_glottolog_phylo_geo(glottolog_version = "4.3"))
}
