% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{get_glottolog_trees}
\alias{get_glottolog_trees}
\title{Glottolog trees by version}
\usage{
get_glottolog_trees(family, glottolog_version)
}
\arguments{
\item{family}{A character vector. Elements are names of glottolog families
whose trees are to be returned. If \code{family} is left unspecified, all
trees are returned.}

\item{glottolog_version}{A character string, specifying which glottolog
version to use. Currently available options are \code{'4.3'} and
\code{'4.4'}. If no value is specified then the newest available version is
used.}
}
\value{
A \code{phylo} object containing one glottolog tree, or a
  \code{multiPhylo} object containing multiple glottolog trees.
}
\description{
Returns a multiPhylo object containing all, or a requested subset, of the
glottolog trees.
}
\details{
By default, trees are returned from the most recent version of glottolog.
Alternatively, an older version of glottolog can be specified.
}
\examples{

library(ape)
tree_totonacan <- get_glottolog_trees(family = "Totonacan")
tree_totonacan_v4.3 <- get_glottolog_trees("Totonacan", "4.3")
plot(tree_totonacan)
plot(tree_totonacan_v4.3)
trees <- get_glottolog_trees(family = c("Caddoan", "Tangkic"))
plot(trees[[1]])
plot(trees[[2]])
}
