% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottolog_geography}
\alias{glottolog_geography}
\alias{glottolog_geography_v4.0}
\alias{glottolog_geography_v4.1}
\alias{glottolog_geography_v4.2}
\alias{glottolog_geography_v4.3}
\alias{glottolog_geography_v4.4}
\alias{glottolog_geography_v4.5}
\alias{glottolog_geography_v4.6}
\title{Geographical data from glottolog, versions 4.0 - 4.6}
\format{
A dataframe:
\describe{
  \item{glottocode}{glottocode of the lect}
  \item{name}{name of the lect}
  \item{isocodes}{the ISO-639-3 code of the lect}
  \item{level}{"language" or "dialect"}
  \item{macroarea}{glottolog's geographical macroarea}
  \item{latitude}{}
  \item{longitude}{}
}

v4.0: A dataframe of 20,049 rows

v4.1: A dataframe of 20,290 rows

v4.2: A dataframe of 20,752 rows

v4.3: A dataframe of 20,930 rows

v4.4: A dataframe of 21,329 rows

v4.5: A dataframe of 21,508 rows

v4.6: A dataframe of 21,652 rows
}
\source{
\url{https://glottolog.org/meta/downloads}
}
\usage{
glottolog_geography_v4.0

glottolog_geography_v4.1

glottolog_geography_v4.2

glottolog_geography_v4.3

glottolog_geography_v4.4

glottolog_geography_v4.5

glottolog_geography_v4.6
}
\description{
Datasets of geographical information about the languages in glottolog.
}
\keyword{datasets}
