% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottolog_geography_v4.4}
\alias{glottolog_geography_v4.4}
\title{Geographical data from glottolog v4.4}
\format{
A dataframe:
\describe{
  \item{glottocode}{glottocode of the lect}
  \item{name}{name of the lect}
  \item{isocodes}{the ISO-639-3 code of the lect}
  \item{level}{"language" or "dialect"}
  \item{macroarea}{glottolog's geographical macroarea}
  \item{latitude}{}
  \item{longitude}{}
}
}
\source{
\url{https://glottolog.org/meta/downloads}
}
\usage{
glottolog_geography_v4.4
}
\description{
A dataset of geographical information about the languages in glottolog.
}
\keyword{datasets}
