% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottolog_trees}
\alias{glottolog_trees}
\alias{glottolog_trees_v4.0}
\alias{glottolog_trees_v4.1}
\alias{glottolog_trees_v4.2}
\alias{glottolog_trees_v4.3}
\alias{glottolog_trees_v4.4}
\alias{glottolog_trees_v4.5}
\alias{glottolog_trees_v4.6}
\alias{glottolog_trees_v4.7}
\alias{glottolog_trees_v4.8}
\title{Trees from glottolog, versions 4.0 - 4.8}
\format{
v4.0: A multiPhylo object of 421 trees

v4.1: A multiPhylo object of 421 trees

v4.2: A multiPhylo object of 422 trees

v4.3: A multiPhylo object of 418 trees

v4.4: A multiPhylo object of 420 trees

v4.5: A multiPhylo object of 420 trees

v4.6: A multiPhylo object of 420 trees

v4.7: A multiPhylo object of 422 trees

v4.8: A multiPhylo object of 422 trees
}
\source{
\url{https://glottolog.org/meta/downloads}
}
\usage{
glottolog_trees_v4.0

glottolog_trees_v4.1

glottolog_trees_v4.2

glottolog_trees_v4.3

glottolog_trees_v4.4

glottolog_trees_v4.5

glottolog_trees_v4.6

glottolog_trees_v4.7

glottolog_trees_v4.8
}
\description{
multiPhylo objects, which provide a representation of the phylogenetic
relationships of the languages in glottolog.
}
\keyword{datasets}
