% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glottolog_trees_v4.4}
\alias{glottolog_trees_v4.4}
\title{Trees from glottolog v4.4}
\format{
A multiPhylo object
}
\source{
\url{https://glottolog.org/meta/downloads}
}
\usage{
glottolog_trees_v4.4
}
\description{
A multiPhylo object which provides a representation of the phylogenetic
relationships of the languages in glottolog.
}
\keyword{datasets}
