% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{keep_as_tip}
\alias{keep_as_tip}
\title{Keep tips and convert nodes to tips}
\usage{
keep_as_tip(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character vector containing tip and node labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
Nominate which tips and nodes are to be kept as tips. Others tips are
removed. If any element of \code{label} is both a node label and tip label,
it will be treated as referring to the tip, not the node.
}
\examples{

library(ape)

tree <- abridge_labels(get_glottolog_trees("Tangkic"))
plot(tree)
nodelabels(tree$node.label)

tree2 <- keep_as_tip(tree, c("lard1243", "kang1283", "nyan1300", "gang1267"))
plot(tree2)
nodelabels(tree2$node.label)
}
