% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{move_node}
\alias{move_node}
\title{Move a node}
\usage{
move_node(phy, label, parent_label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character string containing the label of the node to move.}

\item{parent_label}{A character string containing the label of the new parent
node.}
}
\value{
A phylo object containing the modified tree.
}
\description{
Move one node to a position dominated by a new parent node.
}
\details{
The branch length above the moved node remains unchanged.

If moving the node would result in any other node(s) having no descendant 
tips, then those other nodes are removed.
}
\examples{

library(ape)

tree <- abridge_labels(get_glottolog_trees("LeftMay"))
tree <- ultrametricize(rescale_branches_exp(tree))
plot(tree)
nodelabels(tree$node.label)
tree2 <- move_node(tree, "iter1240", parent_label = "left1242")
plot(tree2)
nodelabels(tree2$node.label)
}
