% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{move_tip}
\alias{move_tip}
\title{Move a tip}
\usage{
move_tip(phy, label, parent_label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character string containing the tip label.}

\item{parent_label}{A character string containing the label of the parent
node.}
}
\value{
A phylo object containing the modified tree.
}
\description{
Move one tip to a new parent node.
}
\details{
In the tip's new position, the length of the branch above it is the same
as the longest branch above any new sister of the tip.

If moving the tip would result in any node(s) having no descendant tips, then
those nodes are removed.
}
\examples{

library(ape)

tree <- abridge_labels(get_glottolog_trees("LeftMay"))
tree <- ultrametricize(rescale_branches_exp(tree))
plot(tree)
nodelabels(tree$node.label)
tree2 <- move_tip(tree, "amap1240", parent_label = "left1242")
plot(tree2)
nodelabels(tree2$node.label)
}
