% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{nonbranching_nodes}
\alias{nonbranching_nodes}
\title{Find nodes with a single child}
\usage{
nonbranching_nodes(phy)
}
\arguments{
\item{phy}{A phylo object}
}
\value{
A vector of node labels
}
\description{
Find nodes with a single child
}
