% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_glotto}
\alias{plot_glotto}
\title{Plot a tree in downward, linguistic style}
\usage{
plot_glotto(
  phy,
  nodelabels = TRUE,
  extra_width = 0,
  extra_height = 0,
  extra_offset = 0,
  ...
)
}
\arguments{
\item{phy}{A phylo object.}

\item{nodelabels}{A logical, whether to plot node labels.}

\item{extra_width}{A numeric, extra width to add beyond the default.}

\item{extra_height}{A numeric, extra height to add beyond the default.}

\item{extra_offset}{A numeric, extra offset to add beyond the default.}

\item{...}{Additional arguments passed to plot().}
}
\description{
An attempt is made to choose a reasonable plot width, height and offset of
the labels from the tree's tips. If the choices are not satisfactory,
manipulate them using positive or negative values of \code{extra_width},
\code{extra_height} and \code{extra_offset}.
}
\examples{

library(ape)

tree <- abridge_labels(get_glottolog_trees("Tangkic"))
plot_glotto(tree)
tree2 <- rescale_deepest_branches(tree, 7)
plot_glotto(tree2)
plot_glotto(tree2, srt = 90, adj = 0.5)
}
