% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{relabel_node}
\alias{relabel_node}
\title{Change node labels}
\usage{
relabel_node(phy, label, new_label)
}
\arguments{
\item{phy}{A phylo object.}

\item{label}{A vector of strings, the node labels to be replaced.}

\item{new_label}{A vector of strings, the same length as \code{label}. The
corresponding replacement labels.}
}
\description{
Replaces existing tip labels with new labels.
}
\examples{

library(ape)

tree <- relabel_with_names(get_glottolog_trees("Tangkic"))
tree2 <- apply_duplicate_suffixes(clone_tip(tree, "Yangkaal", subgroup = TRUE))
plot(tree2)
tree3 <- relabel_tip(tree2, c("Kangkalita", "Yangkaal-1", "Yangkaal-2"), 
                     c("Gangalidda", "Yangkaal", "Yangarella"))
plot(tree3)
nodelabels(tree3$node.label)
tree4 <- relabel_node(tree3, c("Yangkaal", "Kayardild-Yangkaal"),
                      c("Yangkaralda", "Kayardild-Yangkaralda"))
plot(tree4)
nodelabels(tree4$node.label)
}
