% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{relabel_tip}
\alias{relabel_tip}
\title{Change tip labels}
\usage{
relabel_tip(phy, label, new_label)
}
\arguments{
\item{phy}{A phylo object.}

\item{label}{A vector of strings, the tip labels to be replaced.}

\item{new_label}{A vector of strings, the same length as \code{label}. The
corresponding replacement labels.}
}
\description{
Replaces existing tip labels with new labels.
}
\examples{

library(ape)

tree <- relabel_with_names(get_glottolog_trees("Tangkic"))
plot(tree)
tree2 <- relabel_tip(tree, "Kangkalita", "Gangalidda")
plot(tree2)

tree3 <- clone_tip(tree, "Yangkaal", subgroup = TRUE)
plot(tree3)
tree4 <- apply_duplicate_suffixes(tree3)
plot(tree4)
tree5 <- relabel_tip(tree4, c("Kangkalita", "Yangkaal-1", "Yangkaal-2"), 
                     c("Gangalidda", "Yangkaal", "Yangarella"))
plot(tree5)                
}
