% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{relabel_with_names}
\alias{relabel_with_names}
\title{Change labels from glottocodes to names}
\usage{
relabel_with_names(phy, glottolog_version)
}
\arguments{
\item{phy}{A phylo object, the tree to manipulate.}

\item{glottolog_version}{A character string, specifying which glottolog
version to use. Currently available options are \code{'4.0'} through to
\code{'4.8'}. If no value is specified then the newest available version is
used.}
}
\value{
A phylo object, the manipulated tree.
}
\description{
Looks up glottolog language names corresponding to glottocodes and replaces
tip and node labels, which contain glottocodes, with the appropriate names.
}
\details{
Also recognizes and retains duplicate suffixes, i.e., a hyphen followed by
one or more numerals at the end of the string (see
\link{apply_duplicate_suffixes}).

Labels without glottocodes are left unchanged and a warning is given. The
version of glottolog to use for look-up can be controlled with
\code{glottolog_version}.
}
\examples{

library(ape)

# Replace full glottolog labels with names
tree <- get_glottolog_trees("Kresh-Aja")
plot(tree)
nodelabels(tree$node.label)
tree2 <- relabel_with_names(tree)
plot(tree2)
nodelabels(tree2$node.label)

# Replace abridged labels with names
tree3 <- abridge_labels(tree)
plot(tree3)
nodelabels(tree3$node.label)
tree4 <- relabel_with_names(tree3)
plot(tree4)
nodelabels(tree4$node.label)

# Use names from earlier glottolog version:
tree5 <- relabel_with_names(tree, glottolog_version = "4.3")
plot(tree5)
nodelabels(tree5$node.label)
}
