% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{remove_clade}
\alias{remove_clade}
\title{Remove clades}
\usage{
remove_clade(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character vector containing node labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
From a tree, remove clacdes, identified by the labels of their deepest node.
}
