% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{remove_tip}
\alias{remove_tip}
\title{Remove tips}
\usage{
remove_tip(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character vector containing tip labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
From a tree, select which tips are to be removed.
}
\examples{

library(ape)

tree <- abridge_labels(get_glottolog_trees("Tangkic"))
plot(tree)
nodelabels(tree$node.label)

tree2 <- remove_tip(tree, c("kang1283", "kaya1319"))
plot(tree2)
nodelabels(tree2$node.label)

## How to remove all tips below a node, but keep the node as a tip

# By convention, when all tips below a node are removed, the node is removed.
# Here we remove nyan1300 and kaya1319. Node kaya1318 is removed too:
unwanted_tips <- c("nyan1300", "kaya1319")
tree3 <- remove_tip(tree5a, label = unwanted_tips)
plot(tree3)
nodelabels(tree5$node.label)

# step 1 - add a new tip with the same name
tree4a <- add_tip(tree, "kaya1318", parent_node = "sout2758")
plot(tree4)
nodelabels(tree5a$node.label)

# step 2 - remove the unwanted tips and the original node
tree4b <- remove_tip(tree4a, label = unwanted_tips)
plot(tree4b)
nodelabels(tree4b$node.label)
}
