% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{rescale_branches}
\alias{rescale_branches}
\title{Set all branch length to 1}
\usage{
rescale_branches(phy, length = 1)
}
\arguments{
\item{phy}{A phlyo object, the tree to manipulate.}

\item{length}{A numeric stating the branch length.}
}
\value{
A phlyo object, the manipulated tree.
}
\description{
Sets all branch lengths in a tree to the same length.
}
\details{
By default, sets all branch lengths to 1.
}
\examples{

library(ape)
tree <- abridge_labels(get_glottolog_trees("Tangkic"))
tree2 <- clone_tip(tree, "nyan1300", n = 2, subgroup = TRUE)
plot(tree2)
nodelabels(tree2$node.label)
tree3 <- rescale_branches(tree2)
plot(tree3)
nodelabels(tree3$node.label)
}
