% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{rescale_branches_constrained}
\alias{rescale_branches_constrained}
\title{Exponentialize branch lengths using constraints}
\usage{
rescale_branches_constrained(phy, tree_height, shortest_length)
}
\arguments{
\item{phy}{A phylo object, the tree to manipulate.}

\item{tree_height}{A positive numeric, the desired height of the whole tree.}

\item{shortest_length}{A positive numeric, the desired length of the shortest
branches, i.e., the branches farthest from the root.}
}
\value{
A phylo object, the manipulated tree.
}
\description{
Sets the branches furthest from the root to \code{shortest_length} and then
scales other branches exponentially in such a way that the total tree has a
height of \code{tree_height}.
}
\examples{

library(ape)
tree <- abridge_labels(get_glottolog_trees("Siouan"))
plot_glotto(tree)
tree2 <- rescale_branches_constrained(tree, tree_height = 2000, shortest_length = 200)
plot_glotto(tree2)
tree3 <- rescale_branches_constrained(tree, tree_height = 2000, shortest_length = 100)
plot_glotto(tree3)
}
