% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{rescale_branches_exp}
\alias{rescale_branches_exp}
\title{Exponentialize branch lengths}
\usage{
rescale_branches_exp(phy, length = 1)
}
\arguments{
\item{phy}{A phlyo object, the tree to manipulate.}

\item{length}{A positive numeric, a multiplier for the exponential branch
lengths 1/2, 1/4, 1/8...}
}
\value{
A phlyo object, the manipulated tree.
}
\description{
Sets the deepest branches to length 1/2, the next deepest to 1/4, the next to
1/8, etc., all multiplied by the parameter \code{length}.
}
\examples{

library(ape)
tree <- abridge_labels(get_glottolog_trees("Siouan"))
plot(tree)
nodelabels(tree$node.label)
tree2 <- rescale_branches_exp(tree)
plot(tree2)
nodelabels(tree2$node.label)
}
