% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_lengths.R
\name{ultrametricize}
\alias{ultrametricize}
\title{Ultrametricize tree by stretching final edges}
\usage{
ultrametricize(phy)
}
\arguments{
\item{phy}{A phlyo object, the tree to manipulate.}
}
\value{
A phlyo object, the manipulated tree.
}
\description{
Alters branches ending in a tip in such a way that all tips are equidistant
from the root. Does this by lengthening branches above all but the existing,
most-distance tip(s).
}
\examples{

library(ape)
tree <- rescale_branches_exp(abridge_labels(get_glottolog_trees("Siouan")))
plot(tree)
nodelabels(tree$node.label)
tree2 <- ultrametricize(tree)
plot(tree2)
nodelabels(tree2$node.label)
}
