% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{which_tree}
\alias{which_tree}
\title{Tree numbers of glottolog families}
\usage{
which_tree(family, glottolog_version)
}
\arguments{
\item{family}{A character vector. Elements are names of glottolog families
whose trees are to be returned. If \code{family} is left unspecified, all
trees are returned.}

\item{glottolog_version}{A character string, specifying which glottolog
version to use. Currently available options are \code{'4.3'} and
\code{'4.4'}. If no value is specified then the newest available version is
used.}
}
\value{
A named vector of integers, giving the tree numbers and the
  family names as the vector names.
}
\description{
Returns the tree number of one or more glottolog families.
}
\examples{
which_tree("Caddoan")
which_tree(c("Caddoan", "Tangkic"), glottolog_version = "4.3")
# If some family names are unrecognized, a warning is issued
which_tree(c("Caddoan", "Zzz"), glottolog_version = "4.4")
\dontrun{
# If no family names are recognized, an error results
which_tree()
which_tree("Zzz")
}
}
