% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{clone_tip}
\alias{clone_tip}
\title{Clone tips}
\usage{
clone_tip(phy, label, n = 1, subgroup = FALSE)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character vector containing tip labels.}

\item{n}{A numeric vector. The number of clones to make.}

\item{subgroup}{A logical. Whether to create a subgroup containing the new
clones and their original.}
}
\value{
A phylo object containing the modified tree.
}
\description{
Clones tips as sisters of the original. Optionally, places the new clones and
the original in their own subgroup, in which case the node for the new
subgroup is assigned the same label as the original tip.
}
\examples{

library(ape)

tree <- 
  rescale_branches_exp(abridge_labels(get_glottolog_trees("Tangkic")))
plot(tree)
nodelabels(tree$node.label)
tree2 <- clone_tip(tree, "nyan1300")
plot(tree2)
nodelabels(tree2$node.label)

tree3 <- clone_tip(tree, "nyan1300", subgroup = TRUE)
plot(tree3)
nodelabels(tree3$node.label)
# Add suffixes to labels, to keep all labels distinct
tree3a <- apply_duplicate_suffixes(tree3)
plot(tree3a)
nodelabels(tree3a$node.label)

tree4 <- clone_tip(tree, "lard1243", n = 3)
plot(tree4)
nodelabels(tree4$node.label)

tree5 <- clone_tip(tree, "lard1243", n = 3, subgroup = TRUE)
plot(tree5)
nodelabels(tree5$node.label)

tree6 <- clone_tip(tree, c("lard1243", "nyan1300"), n = 2, subgroup = TRUE)
plot(tree6)
nodelabels(tree6$node.label)
tree6a <- apply_duplicate_suffixes(tree6)
plot(tree6a)
nodelabels(tree6a$node.label)

\dontrun{
# Returns error if any element of `label` is not in `phy`
tree7 <-  clone_tip(tree, c("lard1243", "xxxx1234"))
}
}
