% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{remove_tip}
\alias{remove_tip}
\title{Remove tips}
\usage{
remove_tip(phy, label)
}
\arguments{
\item{phy}{A phylo object. The tree to manipulate.}

\item{label}{A character vector containing tip labels.}
}
\value{
A phylo object containing the modified tree.
}
\description{
From a tree, remove tips identified by their node labels.
}
\examples{

library(ape)

tree <- abridge_labels(get_glottolog_trees("Tangkic"))
plot_glotto(tree)

tree2 <- remove_tip(tree, c("kang1283", "kaya1319"))
plot_glotto(tree2)
}
